Fix invoke with void function argument
Summary:
Invokes with a void function argument were not compiling because the void wrapper was not set and because the ref wrapper was on. This change turns on the void wrapper by inferring as Some instead of Any and clears the ref wrapper when the void wrapper is on.

Reviewed By: myang

Test Plan:
new unit test

Revert: OK

DiffCamp Revision: 89748

git-svn-id: http://svn.facebook.net/hiphop/fbomb/branches/hphp-dev/fbcode/hphp@21757 2248de34-8caa-4a3c-bc55-5e52d9d7b73a